# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CreateQuestionRequest:
    properties:
      category_id:
        type: integer
      correct_answer:
        enum:
        - A
        - B
        - C
        - D
        type: string
      option_a:
        type: string
      option_b:
        type: string
      option_c:
        type: string
      option_d:
        type: string
      question_text:
        type: string
      weight:
        minimum: 1
        type: integer
    required:
    - correct_answer
    - option_a
    - option_b
    - option_c
    - option_d
    - question_text
    - weight
    type: object
...
  domain.EnrollEventRequest:
    properties:
      event_id:
        type: string
    required:
    - event_id
    type: object
...
  domain.Question:
    properties:
      category_id:
        type: integer
      correct_answer:
        type: string
      created_at:
        type: string
      id:
        type: string
      option_a:
        type: string
      option_b:
        type: string
      option_c:
        type: string
      option_d:
        type: string
      question_text:
        type: string
      weight:
        type: integer
    type: object
...
  domain.UpcomingEvent:
    properties:
      duration_minutes:
        type: integer
      end_time:
        type: string
      id:
        type: string
      name:
        type: string
      passing_grade:
        type: number
      start_time:
        type: string
    type: object
...
  domain.UpdateQuestionRequest:
    properties:
      category_id:
        type: integer
      correct_answer:
        enum:
        - A
        - B
        - C
        - D
        type: string
      option_a:
        type: string
      option_b:
        type: string
      option_c:
        type: string
      option_d:
        type: string
      question_text:
        type: string
      weight:
        minimum: 1
        type: integer
    required:
    - correct_answer
    - option_a
    - option_b
    - option_c
    - option_d
    - question_text
    - weight
    type: object
...
  domain.UserApproval:
    properties:
      approval_id:
        type: string
      completed_at:
        type: string
      duration_minutes:
        type: integer
      end_time:
        type: string
      event_id:
        type: string
      is_completed:
        type: boolean
      is_passed:
        type: boolean
      name:
        type: string
      passing_grade:
        type: number
      score:
        type: number
      start_time:
        type: string
      started_at:
        type: string
      status:
        description: pending, approved, revoked
        type: string
    type: object
...
  response.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  response.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Operasi berhasil
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Dev Team Pramuka CAT
    url: https://github.com/odealidj/pramuka-CAT
  description: REST API untuk platform Computer Assisted Test (CAT) Pramuka.
  termsOfService: http://swagger.io/terms/
  title: Pramuka CAT API
  version: "1.0"
paths:
...
  /admin/questions:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Question'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Soal
      tags:
      - Admin - Bank Soal
    post:
      consumes:
      - application/json
      parameters:
      - description: Data Soal
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Question'
              type: object
      security:
      - BearerAuth: []
      summary: Buat Soal Baru
      tags:
      - Admin - Bank Soal
  /admin/questions/{id}:
    delete:
      parameters:
      - description: UUID Soal
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Hapus Soal
      tags:
      - Admin - Bank Soal
    get:
      parameters:
      - description: UUID Soal
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Question'
              type: object
      security:
      - BearerAuth: []
      summary: Detail Soal
      tags:
      - Admin - Bank Soal
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID Soal
        in: path
        name: id
        required: true
        type: string
      - description: Data Soal Baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Question'
              type: object
      security:
      - BearerAuth: []
      summary: Update Soal
      tags:
      - Admin - Bank Soal
...
  /protected/exams/enroll:
    post:
      consumes:
      - application/json
      description: 'Peserta mendaftar ke suatu event ujian (status awal: pending)'
      parameters:
      - description: ID Event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.EnrollEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Daftar ke Event Ujian
      tags:
      - Peserta - Ujian
  /protected/exams/my-exams:
    get:
      description: Peserta melihat daftar pendaftaran & nilai ujian miliknya
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.UserApproval'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Riwayat Ujian Saya
      tags:
      - Peserta - Ujian
  /protected/exams/upcoming:
    get:
      description: Peserta melihat event ujian yang belum berakhir
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.UpcomingEvent'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Event Mendatang
      tags:
      - Peserta - Ujian
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"